+++ src/components/organisms/VendingMachine/VendingMachine.stories.js
