+++ src/components/organisms/VendingMachine/VendingMachine.style.js
